requirements.txt
aniso8601==10.0.1
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
clickclick==20.10.2
dnspython==2.8.0
dotenv==0.9.9
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
Flask-RESTful==0.3.10
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==4.0.1
packaging==25.0
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
swagger_ui_bundle==1.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
